Fixed NPE if default is not found for a folder.

+++ src/org/opencms/staticexport/A_CmsStaticExportHandler.java
@@ -383,16 +383,18 @@ public abstract class A_CmsStaticExportHandler implements I_CmsStaticExportHandl
                     // check if the file is the default file of the folder
                     try {
                         CmsResource defaultFile = cms.readDefaultFile(CmsResource.getFolderPath(vfsName));
-                        String defaultfilePath = cms.getRequestContext().removeSiteRoot(defaultFile.getRootPath());
-                        if (vfsName.equals(defaultfilePath)) {
-                            // this is the default file, remove it additionally if present
-                            String rfsNameDefault = CmsResource.getFolderPath(rfsName)
-                                + CmsStaticExportManager.EXPORT_DEFAULT_FILE;
-                            String rfsExportFileName = CmsFileUtil.normalizePath(OpenCms.getStaticExportManager().getExportPath(
-                                vfsName)
-                                + rfsNameDefault.substring(OpenCms.getStaticExportManager().getRfsPrefix(vfsName).length()));
+                        if (defaultFile != null) {
+                            String defaultfilePath = cms.getRequestContext().removeSiteRoot(defaultFile.getRootPath());
+                            if (vfsName.equals(defaultfilePath)) {
+                                // this is the default file, remove it additionally if present
+                                String rfsNameDefault = CmsResource.getFolderPath(rfsName)
+                                    + CmsStaticExportManager.EXPORT_DEFAULT_FILE;
+                                String rfsExportFileName = CmsFileUtil.normalizePath(OpenCms.getStaticExportManager().getExportPath(
+                                    vfsName)
+                                    + rfsNameDefault.substring(OpenCms.getStaticExportManager().getRfsPrefix(vfsName).length()));
 
-                            purgeFile(rfsExportFileName, vfsName);
+                                purgeFile(rfsExportFileName, vfsName);
+                            }
                         }
                     } catch (CmsException e) {
                         // failed to determine default file
